requirements.txt
todoist-python==8.1.3
python-gitlab==2.6.0
rich==9.11.1
toml==0.10.2